requirements.txt
# Core dependencies with EXACT versions for reproducibility
torch==2.1.0
torch-geometric==2.4.0
torch-scatter==2.1.2
torch-sparse==0.6.18
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
pyyaml==6.0.1
matplotlib==3.7.2
seaborn==0.12.2

# Optional but recommended
tqdm==4.66.1
